Use enum for Game.PlayMode

+++ API/Repositories/Implementations/ApiMatchRepository.cs
@@ -282,7 +282,7 @@ public class ApiMatchRepository(
             StartTime = osuApiGame.StartTime,
             EndTime = osuApiGame.EndTime,
             BeatmapId = beatmapIdResult,
-            PlayMode = (int)osuApiGame.Ruleset,
+            PlayMode = osuApiGame.Ruleset,
             ScoringType = (int)osuApiGame.ScoringType,
             TeamType = (int)osuApiGame.TeamType,
             Mods = osuApiGame.Mods
